fix: Restore intake source and split backlog
Automation-created work orders without an origin
showed as Created manually. The backlog create
specs crashed because the card-actions mock did
not include dispatchingOrderIds.

Keep the automation fallback for those orders.
Move BacklogColumn out of LinesPage so the file
does not grow further.
